# Introduce CUDA Device Assertions Infrastructure