… c10/cuda/CUDADeviceAssertionHost.h
… c10/cuda/test/impl/CUDAAssertionsTest_1_var_test.cu
… c10/cuda/test/impl/CUDAAssertionsTest_catches_stream.cu
… c10/cuda/test/impl/CUDAAssertionsTest_catches_thread_and_block_and_device.cu
… c10/cuda/test/impl/CUDAAssertionsTest_from_2_processes.cu
… c10/cuda/test/impl/CUDAAssertionsTest_multiple_writes_from_blocks_and_threads.cu
… c10/cuda/test/impl/CUDAAssertionsTest_multiple_writes_from_multiple_blocks.cu
… c10/cuda/test/impl/CUDAAssertionsTest_multiple_writes_from_same_block.cu