@soumith 2022-09-08
this is pretty cool!