@ezyang 2022-09-27
added @andreaskoepf to the reviewer list, he did an original version of this idea at https://github.com/pytorch/pytorch/pull/26776 ; maybe this time we can actually land it!