@andreaskoepf 2022-09-27
@r-barnes, great to see someone working on cuda asserts again!

When I worked on the graceful-assert feature in https://github.com/pytorch/pytorch/pull/26776 I noticed negative performance impacts as soon as (too many) string-literals appeared in the kernel (either the assert message or `__FILE__` macros), e.g. see my comments  https://github.com/pytorch/pytorch/pull/26776#issuecomment-537575482 and https://github.com/pytorch/pytorch/pull/26776#issuecomment-537854481. I would therefore recommend to run some micro-benchmarks to measure and compare performance of realistic kernels with and without the new cuda-assert infrastructure. In the best case the cuda compiler has been improved during the last years or what I noticed back then simply does not apply to the new approach.

General question: Are there now ways to recover from sticky cuda errors (e.g. `assert(false)`) without terminating the host-process, e.g. is `cudaDeviceReset(`) now enough?

https://docs.nvidia.com/cuda/cuda-runtime-api/group__CUDART__TYPES.html says:
> cudaErrorAssert = 710
>     An assert triggered in device code during kernel execution. The device cannot be used again. All existing allocations are invalid. To continue using CUDA, the process must be terminated and relaunched. 
> 

but https://docs.nvidia.com/cuda/cuda-c-programming-guide/index.html says:

> Any subsequent host-side synchronization calls made for the same device will return cudaErrorAssert. No more commands can be sent to this device until cudaDeviceReset() is called to reinitialize the device.

One of my main motivations in 2019 was to gracefully handle assert-failures in kernels .. I approached it with graceful-returns from the kernels but proper device recovery would of course be much better.