@r-barnes 2022-09-27
@andreaskoepf - Are there existing benchmarks I can use to test for performance degradation?

Device-side assertions still cannot be recovered from without killing the associated host process and this setup does not attempt to work around that. The main goal here is to capture the failed assertions from the device, associate them with whatever kernel launch triggered the failure, and package all this information neatly in one place to facilitate easier debugging.