@r-barnes 2022-09-29
@andreaskoepf - Thanks! When you say
> The module level output when compiling MultinomialKernel.cu with --resource-usage

I wonder how you did that compilation to get the ptxas info?