@r-barnes 2022-09-29
@andreaskoepf

I've built a benchmark on the branch [here](https://github.com/r-barnes/pytorch/blob/dsa_benchmark/temp/dsa_benchmark.cpp), which I've compiled with
```
reset && time CXX=clang++-10 CC=clang-10 USE_CUDA=1 USE_ROCM=0 TORCH_CUDA_ARCH_LIST="8.0" USE_KINETO=0 BUILD_TEST=0 USE_DISTRIBUTED=0 BUILD_CAFFE2_OPS=0 USE_FBGEMM=0 BUILD_CAFFE2=0 USE_OPENCV=0 USE_QNNPACK=0 USE_XNNPACK=0 DEBUG=0 USE_GOLD_LINKER=1 CCACHE_DISABLE=1 python setup.py develop 
```
A diff between the benchmark and this DSA (device-side assertion) infrastructure branch is [here](https://github.com/r-barnes/pytorch/compare/export-D37621532...r-barnes:pytorch:dsa_benchmark).

This uses the same multinomial code as your benchmark, but I've kicked it up to 1k iterations because the N=10 sample size seemed like it was mostly producing noise.

We get the following results:
```
# Baseline (1000 trials per test)
Test #1 2071088
Test #2 2074427
Test #3 2076612
Test #4 2080450
Test #5 2084301
Test #6 2094876
Test #7 2097009
Test #8 2105669
std=11.5292313707376 μs
avg=2085.554 μs

# With DSA (1000 trials per test)
Test #1 2062365
Test #2 2063063
Test #3 2064451
Test #4 2068551
Test #5 2069857
Test #6 2077091
Test #7 2097987
Test #8 2301278
std=76.63315184849424 μs 
avg=2100.580375 μs
```

The standard deviation is 65μs higher with DSA, but I wouldn't expect anyone to notice that. The time difference between the two, at 15μs also seems pretty small.