@ezyang 2022-09-30
What CUDA version is the benchmark on?