@r-barnes 2022-09-30
@ezyang - This is CUDA v11.6.112 (I've edited the description above with this info as well).