@ezyang 2022-09-30
It's possible that @andreaskopf's cited slowdown was specific to CUDA 10. If so, happy day; since we're dropping 10.2 support. If you want to be totally sure, test with the earliest CUDA version we support (sorry, I actually don't remember which point release it is...)