@ezyang 2022-09-30
11.3, according to @janeyx99 !

I wonder if the fact that we no longer need out of line string literals means we can also stop trashing the CUDA context. Maybe we can do it as follow up.