@janeyx99 2022-09-30
@r-barnes @ezyang it appears 11.3 was a misleading answer! the stable torch 1.12 still supports 11.3 but our nightly binaries have 11.6 as the lowest cuda version we support.