@andreaskoepf 2022-09-30
Probably the macro as it is currently proposed `CUDA_KERNEL_ASSERT2(condition)` is fine .. developers would have to write something like `&& "msg"` to add string messages and probably most of the time will not naturally do this. For debug purposes it is already great to get a string-representation of the failing condition together with the source-reference and at least in the multinomial test that seemed not to harm performance. Overall 👍 from my side.