@r-barnes 2022-09-30
@andreaskoepf - Apologies, I've added print statements to both the CUDA and CPU multinomial code and it doesn't look as though my benchmark is using either. I'm not sure what's going wrong, but it looks like the code is getting its multinomial implementation from some place other than the source I modified.