@andreaskoepf 2022-09-30
> I wonder how you did that compilation to get the ptxas info?

I probably added [--resource-usage (-res-usage)](https://docs.nvidia.com/cuda/cuda-compiler-driver-nvcc/index.html#generic-tool-options-resource-usage) or `--ptxas-options=-v` to the nvcc args.