@ezyang 2022-10-17
Sorry, was really behind on reviewing this.

At a high level, this looks pretty reasonable. I am primarily worried about the performance implications of the patch. It would be good to hook this up to at least one real kernel in this PR, and then microbenchmark the impact, including in multithreaded settings. I'd also love to know if we can avoid trashing the cuda context, this matters a lot to folks and if inline strings are OK for perf, I don't think it's that much more work to do.