@ngimel 2022-10-17
@ezyang wdym not trashing cuda context?