@ezyang 2022-10-17
If you trigger a cuda assert in a notebook you don't have to restart the notebook