@ezyang 2022-10-17
ok fine, suggestion retracted