@r-barnes 2022-10-25
@ezyang: We can save the CUDA context, but it will be difficult to truly recover from a device-side assertion failure without rolling back to some known checkpoint. CUDA kernels run asynchronously, which means _many_ kernels can be scheduled to run without intervening host operations that could detect device-side failures. Once a device-side failure has been detected it would still be difficult to determine which kernels ran before the failure occurred. 

`C10_CUDA_CHECK(cudaDeviceSynchronize())` is the only call we can make that's immune to this issue since`cudaDeviceSynchronize` blcoks until all kernels have run and then `C10_CUDA_CHECK` can check for failures. If this succeeds, then the code is at a known good point. If this fails, then some tensors have been garbagified since the last such call.