@ezyang 2022-10-25
I think the most practical way to figure out if the tensors are valid or not would to be have some globally increasing atomic sequence number that we assign to kernel launches. When an assert fails we note which sequence number failed, and then all tensors after that sequence number are now poisoned.