@r-barnes 2022-10-25
@ezyang - That counter's already in this PR. But I don't think there's a convenient way to map kernel launches to their associated (Python?) output tensors.