@ezyang 2022-10-25
Yes, we would have to have space to store this on the tensor somehow